Use regions to split class and small refactor

+++ ClangPowerTools/ClangPowerTools/MVVM/Controllers/LlvmController.cs
@@ -10,6 +10,7 @@ namespace ClangPowerTools.MVVM.Controllers
 {
   public class LlvmController
   {
+    #region Members
     public LlvmSettingsModel llvmModel = new LlvmSettingsModel();
     public delegate void SetInstallCommandState();
     public delegate void SetUninstallCommandState();
@@ -23,14 +24,33 @@ namespace ClangPowerTools.MVVM.Controllers
     private const string arguments = @"/C reg delete HKEY_LOCAL_MACHINE\SOFTWARE\WOW6432Node\Microsoft\Windows\CurrentVersion\Uninstall\LLVM /f &";
     private const string processFileName = "cmd.exe";
     private const string processVerb = "runas";
-    private const string uri = @"http://releases.llvm.org";
+    private const string llvmReleasesUri = @"http://releases.llvm.org";
     private const string llvm = "LLVM";
     private const string uninstall = "Uninstall";
 
     private Process process;
     private CancellationTokenSource downloadCancellationToken = new CancellationTokenSource();
     private SettingsPathBuilder settingsPathBuilder = new SettingsPathBuilder();
+    private string bitOperatingSystem = string.Empty;
+    #endregion
 
+
+    #region Public Methods
+    public void DownloadLlvmVersion(string version, DownloadProgressChangedEventHandler method)
+    {
+      CreateVersionFolder(version);
+
+      var executablePath = settingsPathBuilder.GetLlvmExecutablePath(version, llvm + version);
+      var uri = string.Concat(llvmReleasesUri, "/", version, "/", llvm, "-", version, GetOperatingSystemParamaters());
+
+      using (var client = new WebClient())
+      {
+        client.DownloadProgressChanged += method;
+        client.DownloadFileCompleted += DownloadFileCompleted;
+        downloadCancellationToken.Token.Register(client.CancelAsync);
+        client.DownloadFileAsync(new Uri(uri), executablePath);
+      }
+    }
 
     public void InstallLlVmVersion(string version)
     {
@@ -46,7 +66,6 @@ namespace ClangPowerTools.MVVM.Controllers
         process.StartInfo.Verb = processVerb;
         process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
         process.EnableRaisingEvents = true;
-        //TODO clear events or try using
         process.Exited += new EventHandler(InstallProcessExited);
         process.Exited += new EventHandler(InstallFinished);
         process.Start();
@@ -55,54 +74,10 @@ namespace ClangPowerTools.MVVM.Controllers
       catch (Exception e)
       {
         setInstallCommandState();
+        process.Kill();
         DeleteLlvmVersion(llvmModel.Version);
         MessageBox.Show(e.Message, "Installation Failed", MessageBoxButtons.OK, MessageBoxIcon.Error);
       }
-    }
-
-    private void InstallProcessExited(object sender, EventArgs e)
-    {
-      process.Close();
-      var exeName = string.Concat("LLVM", llvmModel.Version, ".exe");
-      DeleteExe(llvmModel.Version, exeName);
-      setInstallCommandState();
-    }
-
-    public void DownloadLlvmVersion(string version, DownloadProgressChangedEventHandler method)
-    {
-      CreateVersionFolder(version);
-
-      //TODO check windows version
-      var executablePath = settingsPathBuilder.GetLlvmExecutablePath(version, llvm + version);
-      var finalUri = string.Concat(uri, "/", version, "/", llvm, "-", version, "-win64.exe");
-
-      using (var client = new WebClient())
-      {
-        client.DownloadProgressChanged += method;
-        client.DownloadFileCompleted += DownloadFileCompleted;
-        downloadCancellationToken.Token.Register(client.CancelAsync);
-        client.DownloadFileAsync(new Uri(finalUri), executablePath);
-      }
-    }
-
-    private void DownloadFileCompleted(object sender, AsyncCompletedEventArgs e)
-    {
-      if (downloadCancellationToken.IsCancellationRequested || llvmModel.DownloadProgress != llvmModel.MaxProgress)
-      {
-        setUninstallCommandState();
-        DeleteLlvmVersion(llvmModel.Version);
-        MessageBox.Show("The download process has stopped.", "Download", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-      }
-      else
-      {
-        llvmModel.IsInstalling = true;
-        llvmModel.IsDownloading = false;
-        InstallLlVmVersion(llvmModel.Version);
-      }
-
-      llvmModel.DownloadProgress = 0;
-      downloadCancellationToken.Dispose();
-      downloadCancellationToken = new CancellationTokenSource();
     }
 
     public void UninstallLlvmVersion(string version)
@@ -133,6 +108,44 @@ namespace ClangPowerTools.MVVM.Controllers
       }
     }
 
+    public bool IsVersionExeOnDisk(string version, string name)
+    {
+      var executablePath = settingsPathBuilder.GetLlvmExecutablePath(version, name);
+      return File.Exists(executablePath);
+    }
+
+    #endregion
+
+
+    #region Private Methods
+    private void InstallProcessExited(object sender, EventArgs e)
+    {
+      process.Close();
+      var exeName = string.Concat(llvm, llvmModel.Version, ".exe");
+      DeleteExe(llvmModel.Version, exeName);
+      setInstallCommandState();
+    }
+
+    private void DownloadFileCompleted(object sender, AsyncCompletedEventArgs e)
+    {
+      if (downloadCancellationToken.IsCancellationRequested || llvmModel.DownloadProgress != llvmModel.MaxProgress)
+      {
+        setUninstallCommandState();
+        DeleteLlvmVersion(llvmModel.Version);
+        MessageBox.Show("The download process has stopped.", "Download", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+      }
+      else
+      {
+        llvmModel.IsInstalling = true;
+        llvmModel.IsDownloading = false;
+        InstallLlVmVersion(llvmModel.Version);
+      }
+
+      llvmModel.DownloadProgress = 0;
+      downloadCancellationToken.Dispose();
+      downloadCancellationToken = new CancellationTokenSource();
+    }
+
     private void UninstallProcessExited(object sender, EventArgs e)
     {
       process.Close();
@@ -140,17 +153,10 @@ namespace ClangPowerTools.MVVM.Controllers
       DeleteLlvmVersion(llvmModel.Version);
     }
 
-    public bool IsVersionExeOnDisk(string version, string name)
-    {
-      var executablePath = settingsPathBuilder.GetLlvmExecutablePath(version, name);
-      return File.Exists(executablePath);
-    }
-
-
     private void DeleteLlvmVersion(string version)
     {
       var path = settingsPathBuilder.GetLlvmPath(version);
-      Directory.Delete(path, true);
+      if(Directory.Exists(path)) Directory.Delete(path, true);
     }
 
     private void CreateVersionFolder(string version)
@@ -164,5 +170,21 @@ namespace ClangPowerTools.MVVM.Controllers
       var path = Path.Combine(settingsPathBuilder.GetLlvmPath(version), exeName);
       File.Delete(path);
     }
+
+    private string GetOperatingSystemParamaters()
+    {
+      if (Environment.Is64BitOperatingSystem)
+      {
+        bitOperatingSystem = "-win64.exe";
+      }
+      else
+      {
+        bitOperatingSystem = "-win32.exe";
+      }
+
+      return bitOperatingSystem;
+    }
+
+    #endregion
   }
 }
